requirements.txt
# Telegram Bot
python-telegram-bot==20.7
python-telegram-bot[webhooks]==20.7

# AI Services
openai==1.12.0
elevenlabs==0.2.27

# Calendar Integration
caldav==1.3.9
icalendar==5.0.11
google-api-python-client==2.108.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0

# HTTP & Async
aiohttp==3.9.1
aiofiles==23.2.1
httpx==0.25.2

# Date/Time
python-dateutil==2.8.2
pytz==2023.3

# Configuration
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Audio Processing
pydub==0.25.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-timeout==2.2.0

# Utilities
loguru==0.7.2
aioresponses==0.7.8
